pyproject.toml
[project]
name = "python-flask-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "cryptography>=46.0.3",
    "flask>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.0",
]
